update cases when files already exists

+++ get_data.sh
@@ -29,16 +29,18 @@ echo $ncase
 #var_list=(FISCCP1_COSP FSDSC FSNSC TREFHT FSNT FSNTC FLUT FLUTC TS FSDS FSNS TGCLDLWP TGCLDIWP T Q)
 #var_new_list=(clisccp rsdscs rsnsc tas rsnt rsntcs rlut rlutcs ts rsds rsns TGCLDLWP TGCLDIWP ta hus)
 
-vars=(FISCCP1_COSP,FSDSC,FSNSC,TREFHT,FSNT,FSNTC,FLUT,FLUTC,TS,T,CLOUD,CLDLIQ,CLDICE,FSDS,FSNS,Q,SOLIN,FSUTOA,FSUTOAC,PSL,PS)
-var_list=(FISCCP1_COSP FSDSC FSNSC TREFHT FSNT FSNTC FLUT FLUTC TS T CLOUD CLDLIQ CLDICE FSDS FSNS Q SOLIN FSUTOA FSUTOAC PSL PS)
-var_new_list=(clisccp rsdscs rsnsc tas rsnt rsntcs rlut rlutcs ts ta CLOUD CLDLIQ CLDICE rsds rsns hus rsdt FSUTOA FSUTOAC psl ps)
+#vars=(FISCCP1_COSP,FSDSC,FSNSC,TREFHT,FSNT,FSNTC,FLUT,FLUTC,TS,T,CLOUD,CLDLIQ,CLDICE,FSDS,FSNS,Q,SOLIN,FSUTOA,FSUTOAC,PSL,PS)
+#var_list=(FISCCP1_COSP FSDSC FSNSC TREFHT FSNT FSNTC FLUT FLUTC TS T CLOUD CLDLIQ CLDICE FSDS FSNS Q SOLIN FSUTOA FSUTOAC PSL PS)
+#var_new_list=(clisccp rsdscs rsnsc tas rsnt rsntcs rlut rlutcs ts ta CLOUD CLDLIQ CLDICE rsds rsns hus rsdt FSUTOA FSUTOAC psl ps)
+
+vars=(FISCCP1_COSP,FSDSC,FSNSC,TREFHT,FSNT,FSNTC,FLUT,FLUTC,TS,T,FSDS,FSNS,Q,SOLIN,FSUTOA,FSUTOAC,PSL,PS)
+var_list=(FISCCP1_COSP FSDSC FSNSC TREFHT FSNT FSNTC FLUT FLUTC TS T FSDS FSNS Q SOLIN FSUTOA FSUTOAC PSL PS)
+var_new_list=(clisccp rsdscs rsnsc tas rsnt rsntcs rlut rlutcs ts ta rsds rsns hus rsdt FSUTOA FSUTOAC psl ps)
 
 
 nvar=${#var_list[@]}
 echo $nvar
 
-#int_year=1
-#end_year=5
 int_mon=1
 end_mon=12
 
@@ -53,7 +55,7 @@ do
 #	outdir=/global/cscratch1/sd/qinyi/E3SM_middata/${run_id[ii]}/
 
     if [ "${run_id[ii]}" == "20200428.DECKv1b_amip1-CFMIP.ne30_oEC.cori-knl-L" ] ; then
-        datadir=${datadir_in1}/${run_id[ii]}/${datadir_in2}h0/
+        datadir=${datadir_in1}/${run_id[ii]}/${datadir_in2}
     else
         datadir=${datadir_in1}/${run_id[ii]}/${datadir_in2}
     fi
@@ -64,14 +66,14 @@ do
 
 	if [ ! -d "${outdir}" ] ; then
 		mkdir -p ${outdir}
-    else
-		# if the directory is not empty, pls don't continue the following codes
-		if [ -z "$(ls -A ${outdir})" ]; then
-			echo "outdir is Empy"
-		else
-			echo "outdir Not Empty"
-			continue
-		fi
+    #else
+	#	# if the directory is not empty, pls don't continue the following codes
+	#	if [ -z "$(ls -A ${outdir})" ]; then
+	#		echo "outdir is Empy"
+	#	else
+	#		echo "outdir Not Empty"
+	#		continue
+	#	fi
 	fi
 	
 	if [ "$regrid" == "True" ] ; then
@@ -86,60 +88,74 @@ do
 				imon_2d=`printf %02d $imon`
 				echo "yr=" ${iyr_4d},"month=" ${imon_2d}
 				file_tmp=`ls $datadir/${run_id[ii]}.cam.h0.${iyr_4d}-${imon_2d}.nc`
-		#		all_file_list="${all_file_list} ${list_tmp}"
 				out_file=${run_id[ii]}.cam.h0.${iyr_4d}-${imon_2d}_regrid.nc
 		
-				#if [ ! -f "$outdir/${out_file}" ] ; then
+				if [ ! -f "$outdir/${out_file}" ] ; then
 					echo "regrid file to", $out_file
 					ncremap -v ${vars} -m ${rgr_map} ${file_tmp} $outdir/${out_file}
-				#fi
+                else
+                    echo "The output file is ready:" $out_file
+				fi
+
+				all_file_list="${all_file_list} ${out_file}"
+
 			done
 		done
-
-		# use 'ncrcat' to get monthly-series variables
-		ncrcat -O $outdir/*.cam.h0.????-??_regrid.nc $outdir/${run_id[ii]}.cam.h0.${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
-        rm $outdir/*.cam.h0.????-??_regrid.nc
-
 	fi
-	
-	echo $outdir
-	in_file=$outdir/${run_id[ii]}.cam.h0.${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
-	
-	cd $outdir
-	
-	for ivar in `seq 0 $[$nvar-1]`
-	do
-		echo ${var_new_list[ivar]}
-		out_file=${var_new_list[ivar]}_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
-		# extract each variable from the input file
-		ncks -O -v ${var_list[ivar]} ${in_file} ${out_file}
-		# change variable name
-        if [ "${var_list[ivar]}" != "${var_new_list[ivar]}" ] ; then
-    		ncrename -v ${var_list[ivar]},${var_new_list[ivar]} ${out_file}
-        fi
-	done
-	
-	# get surface up SW at clear-sky
-    ncks -O -v FSDSC,FSNSC,FSDS,FSNS,SOLIN,FSNT,FSNTC ${in_file} tmp.nc
-	ncap2 -O -s 'rsuscs=FSDSC-FSNSC' tmp.nc rsuscs_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
-    ncks -O -v rsuscs rsuscs_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc rsuscs_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
 
-     # get surface up SW at all-sky
-	ncap2 -O -s 'rsus=FSDS-FSNS' tmp.nc rsus_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
-    ncks -O -v rsus rsus_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc rsus_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+	# use 'ncrcat' to get monthly-series variables
+    cd $outdir
+    checkfile=rsuscs_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
 
-	ncap2 -O -s 'rsut=SOLIN-FSNT' tmp.nc rsut_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
-    ncks -O -v rsut rsut_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc rsut_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+    tmpfile=${run_id[ii]}.cam.h0.${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
 
-	ncap2 -O -s 'rsutcs=SOLIN-FSNTC' tmp.nc rsutcs_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
-    ncks -O -v rsutcs rsutcs_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc rsutcs_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+    if [ ! -f "$outdir/$checkfile" ] ; then
 
-    rm tmp.nc
+        echo "ncrcat all necessary monthly files..."
+		ncrcat -O ${all_file_list} $outdir/${tmpfile}
 
-	echo $outdir
-    rm $outdir/${run_id[ii]}.cam.h0.${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+	    in_file=$outdir/${run_id[ii]}.cam.h0.${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+	    cd $outdir
+	    
+	    for ivar in `seq 0 $[$nvar-1]`
+	    do
+	    	echo ${var_new_list[ivar]}
+	    	out_file=${var_new_list[ivar]}_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+        	# extract each variable from the input file
+        	ncks -O -v ${var_list[ivar]} ${in_file} ${out_file}
+        	# change variable name
+            if [ "${var_list[ivar]}" != "${var_new_list[ivar]}" ] ; then
+        		ncrename -v ${var_list[ivar]},${var_new_list[ivar]} ${out_file}
+            fi
+	    done
+	    
+	    # get surface up SW at clear-sky
+        ncks -O -v FSDSC,FSNSC,FSDS,FSNS,SOLIN,FSNT,FSNTC ${in_file} tmp.nc
+        ncap2 -O -s 'rsuscs=FSDSC-FSNSC' tmp.nc rsuscs_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+        ncks -O -v rsuscs rsuscs_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc rsuscs_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+
+        # get surface up SW at all-sky
+	    ncap2 -O -s 'rsus=FSDS-FSNS' tmp.nc rsus_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+        ncks -O -v rsus rsus_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc rsus_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+
+        # get TOA up SW at all-sky
+        ncap2 -O -s 'rsut=SOLIN-FSNT' tmp.nc rsut_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+        ncks -O -v rsut rsut_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc rsut_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+
+        # get TOA up SW at clear-sky
+        ncap2 -O -s 'rsutcs=SOLIN-FSNTC' tmp.nc rsutcs_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+        ncks -O -v rsutcs rsutcs_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc rsutcs_${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+
+        rm tmp.nc
+        rm $outdir/${run_id[ii]}.cam.h0.${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc
+
+    else
+        echo "All necessary output files are ready."
+        echo `ls *${exp_id[ii]}_${int_year_4d}${int_mon_2d}-${end_year_4d}${end_mon_2d}.nc`
+    fi
 
 done 
 
-
+echo "The final outdir =" $outdir
+echo "Well Done!"
